+++ bin/check-dcos-jobs-health.rb
@@ -52,7 +52,7 @@ class CheckDcosJobsHealth < Sensu::Plugin::Check::CLI
          description: 'URL',
          short: '-u URL',
          long: '--url URL',
-         default: 'http://leader.mesos:5050/tasks'
+         default: 'http://leader.mesos:5050/tasks?order=desc'
 
   option :pattern,
          description: 'Pattern',
@@ -62,32 +62,56 @@ class CheckDcosJobsHealth < Sensu::Plugin::Check::CLI
 
   option :window,
          description: 'Window/history for tasks',
-         short: '-w time',
-         long: '--window time',
-         default: 1000.0000
+         short: '-w time minutes',
+         long: '--window time minutes',
+         default: 15
 
-  option :threshold,
-         description: 'Threshold for a running task - the Window',
-         short: '-t float',
-         long: '--threshold float',
-         default: 200.0000
+  option :max,
+         description: 'Maximum time the running task',
+         short: '-M minutes',
+         long: '--max minutes',
+         default: 12
+
+  option :min,
+         description: 'Minimum time the task should run',
+         short: '-m minutes',
+         long: '--min minutes',
+         default: 1
 
   def run
     t = Time.now.to_f.round(4)
     resource = get_data(config[:url])
     resource['tasks'].each do |unit|
-      if unit['id'] =~ /#{config[:pattern].sub('.', '.*')}/ && unit['statuses'][0]['timestamp'] > t - config[:window].to_f.round(4)
+      if unit['id'] =~ /#{config[:pattern].sub('.', '.*').sub(' ', '|')}/ && unit['statuses'][0]['timestamp'] > t - (60 * config[:window].to_f.round(4))
         if unit['state'] =~ /RUNNING/
-          if t - unit['statuses'][0]['timestamp'] > (config[:window].to_f.round(4) - config[:threshold].to_f.round(4))
+          if t - unit['statuses'][0]['timestamp'] > (60 * config[:max].to_f.round(4))
             message "JOB: #{unit['id']} is taking too long to finish..."
             critical
+          else
+            message "#{unit['id']}"
+            ok
           end
         elsif unit['state'] =~ /FAILED|KILLED/
           message "JOB: #{unit['id']}"
           critical
+        elsif unit['statuses'][1]['timestamp'] - unit['statuses'][0]['timestamp'] < (60 * config[:min].to_f.round(4))
+          message "JOB: #{unit['id']} took less then expected"
+          critical
+        else
+          message "#{unit['id']}"
+          ok
         end
+      elsif unit['id'] =~ /#{config[:pattern].sub('.', '.*').sub(' ', '|')}/ && unit.key?("statuses")
+        if t < (unit['statuses'][0]['timestamp'] + (60 * config[:window].to_f.round(4)) + 60)
+          message "#{unit['id']}"
+          ok
+        else
+          message "JOB: did not run on schedule"
+          critical
+        end
+      else
+        next
       end
     end
-    ok
   end
 end